cf/requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.2
drf-url-filters==0.5.1
gunicorn==19.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
Pygments==2.4.2
pytz==2019.2
sqlparse==0.3.0
voluptuous==0.11.7
whitenoise==4.1.3
